requirements.txt
# API & server
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
httpx>=0.27.0

# Agent & RAG
langgraph>=0.2.0
openai>=2.0.0
sentence-transformers>=2.3.0
faiss-cpu>=1.8.0
numpy>=1.24.0
tiktoken>=0.5.0
